+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add tenant filtering in Alloy config
+- Implement `grafana-organization` controller to update tenant IDs filter on GrafanaOrganization changes.
 
 ### Changed